Refactor workspace endpoints and CbsGeoClient parsing logic
